feat(VehicleStore): add store and display data in list
this commit introduces an addition of a mobx store along with the store provider.
the initial data for the store is hydrated on start via getServerSideProps by making a fetch request to the API where data is stored.
also, one more addition is that the data inside the store is displayed in the list layout where every car model is a list item.