fix(VehicleStore): redefine currentVehicle initial state
we do this because we'll get a complaint from react in the console due to the values being undefined or null, the error is:

Warning: A component is changing a controlled input to be uncontrolled. This is likely caused by the value changing from a defined to undefined, which should not happen. Decide between using a controlled or uncontrolled input element for the lifetime of the component. More info: https://reactjs.org/link/controlled-components